Setup composer and autoload
Keeping things in classes and namespaces is required for proper
organization.

This sets up composer and moves some logic around to files that
we will expand on later.

The "index" will just setup autoload and pass control to a "site
runner" which will eventually choose which page we want to display.

The "server" utils will handle all the google cloud stuff. Some
dev things can only be done in the cloud and it takes a while to
get things uploaded for rapid development.

The "siteRunner" utils will eventually display a few different
possible pages.