Add basic template engine
We're not doing anything fancy here but we still want to separate
the html from the php as much as we can.

This creates a "Page" object that can take in some arbitrary data
and pass it to a "phtml" which should just basically be a html
file with access to some php data.