Abstract Page Object
Now that we've got a single site page printing html with css and js,
we'll want to make multiple pages work.

Currently there's only a Page class, but we can abstract a lot
of the "use php data to print text" logic into a "BasePage" class
and the page specific data, the template and title, into it's own
object.

This should make adding a "new" page to display as simple as
creating a phtml file and setting a title to it.

NOTE: The route logic still needs to be setup but we'll do that
next, after we can support multiple html pages.