Create Basic Template Engine
Now that we can view multiple pages, we'd like to add new pages
easily while still keeping the cool layout we picked.

This breaks up each of the major HTML elements in the layout into
"element" objects linked to a phtml template.

Each object builds off of each nested element which basically allows
for the nav, header, head and footer to be the same but the "article"
section easily changed by any BasePage.

This should allow for several pages all with the same look but
only worrying about a small section when we need to add things.

The about_me.phtml changes will probably help explain a lot of
what happened.

In this end result, we have "Hello World" being printed in /dev
but it in same layout as "about me".