Enable Google Cloud Debugger
The google cloud error handler does some weird things if this isn't
installed, so lets install it.

It's got a lot of cool bells & whistles so there's a lot to be
gained and the code change is minimal.

This should add all the required app/composer/php.ini changes and
enable the debugger as one of the first things in the php logic.

It also does a couple minor fixes
 * Add `use Exception` in DevPage to actually do what we want
 * Add a link to the "handle secrets in yaml" issue

NOTE: It'll be tested by throwing an Exception in /Dev

closes #11