Load Nav data from the DB
The database `page_index` table holds the list of pages we want
to display, so lets use it to generate the sidebar Nav data.

Not much special going on here. Only main concern is to not hit
the database more than needed. We just want all the rows from
a single table for now so we don't need to do much but stick
the data we fetch into a common object, the `BasePage`.