Add "HTML Packet" concept and use for Nav
It's nice to have the object that generates the data we want to
display in HTML separate from the logic that actually outputs the
HTML b/c than we can test things without messing with html.

This creates a simple "Packet" object that all the HTML elements
will have that will calculate & hold all the data we'd want to use
in the element.

The .phtml should only be outputing data and minor control logic.