Create packet for Head Element
Last Packet that we need. Now the html should stay pretty clean,
the data separated & we should be able to test anything right
before it would be displayed.

Seems like a lot of setup but future changes will be much simpler
and additions will have a obvious structure to use.