Add "slug" data and use it in the SiteRunner
We're going to display database values to the user but we didn't
really have a way of know what "page" we were on.

A common way of denoting pages is using a "slug" so we'll do that
for our few pages here.

Each page now requires a "slug" & an entry in the SiteRunner
getAllPages call.

The site runner will now display pages based on the PAGE_SLUG
constant in each of the Page files.