Add some widgets & visual polish to the About Me page
We've got a cool framework but not many examples of what it can do.

This adds some custom HTML, css and js to add some color and
interactivity to the page.

Now some random things on the about-me page will be colorful and
you can click the image to refresh the colors.

Some Lorum Ipsum was also added to fill the page out.

closes #25