Add `page_nav` table and update query to use it
We want to have a separate table for the site navigation so we can
do some fancier things with the nav bar.

This moves the nav data from `page_index` to `page_nav` and updates
the query to just join the tables to get the same row results.

In the end we get the same things but with 2 tables.

part of #37