Add type to page_nav table
I was planning on doing some basic logic where if the pageid was
null, we'd use the custom url, otherwise we'd load the pageid data
but doing these weird logical switch things off of arbitrary values
ends up confusing really fast.

So we'll just add a `type` field that defines some basic behaviors
and then it's more obvious what and why.

Part of #37