Move and rename DB.php
Now that we're going to have more database tables to work with,
having all the logic in one file doesn't seem like a good idea.

I think we're also going to have several files, each one relating
to a table, and they have a specific purpose; interacting with the db.

To that effect:
 * DB.php moved to a name more related to the logic: PDOConnection
 * Namespace moved from Utils to DB
 * References to Utils\DB are now DB\PDOConnection
 * Adds new namespace to compose.json
   * Alphabetizes the includes b/c some order is better than none

Part of #28, #37