Move PageIndex logic out of PDOConnection
Now that we have a file that's just supposed to handle the db
connection, we want to move all the page_index specific logic
out of it.

This creates the PageIndex file and moves all the page_index
data into it.

Specifically:
 * The function we'd expect an array of data from
 * The "offline" data
 * The query we'd want to run to get the array of data

Part of #28
The PDOConnection still handles all the actual connection logic
and there's a kinda hack to only make 1 PDOConnection per session
b/c there's no reason to generate a ton of them.