Add Mini Article & Tag db schema
We want to Mini Articles to the pages but we don't have any tables
to pull Mini Article data from.

This adds the required table definitions and adds some dummy data
that we can use on the site.

In theory, it'll add 5 Mini Articles & 6 tags. Each Mini Article
will have at least 2 tags associated with it.

Closes #27