Build Mini Article Widget HTML
We'd like to build some fancy web code that will allow us to display
a list of text entries with some filtering and sorting of those
entries.

This requires some specific js, css and HTML to do correclty and
we can't rely on the markup parser to give us a HTML stucture we
can use.

So we can create a MiniArticleList Widget which we will use to
genreate all the fancy HTML structures we're gonna need.

NOTE: There's quite a lot going on here but it's difficult to break
up without losing the scope of what we're doing.

A lot of it is just generating HTML strings. The logic is kinda
ugly but a general library isn't much easier to understand.

Closes #29, #30