Break JS into classes
JS gets super messy really fast, especially if we're gonna run
"vanilla" which we should be able to do.

IMO the best way to handle it is to break everything up into
classes based on some theme and try to keep all those classes
at a reasonable size (<1k lines).

I'm still not ready for several files (and don't really want to
support that) so this breaks all the JS logic into separate classes
based on some theme but keeps all the classes in the same file.

Part of #48