Add Filtering by Tag to Mini Article List
We have a section in the MiniArticleList that shows all the possible
tags in the Mini Articles.

This adds the js logic so that when you click on one of the possible
tags, all the mini articles with that tag will be displayed. If the
tag is clicked again, all the filtered mini articles will show up.

closes #48