Add sorting to Mini Article List
There are some dates associated with the mini articles and it
would be cool if they did somthing other than look pretty.

This adds some sort options at the top of the Mini Article List
and updates the JS to handle sorting the Mini Articles by their
"start date"

It's vanilla JS so kinda ugly but I tried to add comments on
the worst of it.

closes #52