Clean up ServerUtils logic
One of the first files we created was Utils\Server and it's
basically answered one question the entire time:
"Are we in the Google Cloud?"

We still mostly only care about this question, but we've started
using it for very different reasons:
 * Loading google debug tools
 * Loading data from the db
 * Loading google secrets

This breaks up all those questions into separate functions. This
helps create a code pattern of being specific with configuration
definitions.

It also allows us to "turn off the db" by toggling a return value, which
is important in keeping costs down: https://github.com/cdcline/demo-website/issues/50

We've also gotten into a bit of a pattern where utils class that don't
expect to be instantiated have `Utils` in their name. We'll try
to keep that pattern going by renaming the file to `ServerUtils`
which is name being used in most of the classes anyways.

Notes:
 * Because JS also has a ServerUtils class, we'll rename it JSServerUtils
  * Yes this isn't needed, but I like strings to be unique for code searches
 * There are still some \Utils classes that don't have Utils in the class name
  * SecretManager is an object so shouldn't have it by our pattern
  * Parser is only static functions but it should probably be a better object
   * Not quite sure what to do with it but it serves it's purpose of keeping "parsing" someplace specific
  * SiteRunner is only static function but will proably be an object soon
   * Will probably be converted when working on: https://github.com/cdcline/demo-website/issues/32

Part of #50