Clean up Mini Article List code and css
The MiniArticleList widget is a lot of html, css and js.

It's confusing to follow when there are 3 pieces tied together and
it's also confusing why we have a "widget" concept in the first place.

This tries to add comments and clean up the Mini Article List code
to be more understandable about what's happening.

Specifically:
 * Added structure and attributes comments to MiniArticleList
 * Added a "widget" trait with a brief description on what they are
 * Renamed several things in the MiniArticleList to be (hopefully) more legibal
 * Adds some css styles to the MiniArticleList so it looks a little cleaner
  * Also sets up css targets for future changes

part of #22, #26