Minor MiniArticle cleanup; add vscode to .gitignore
We'd like the Mini Article tags to be in the same order, no matter
what order the entry tags are.

This adds a sort to the tags array to keep the tags in alphabetical
order. The tags on the Entries themselvs are basically in order of
adding them but that's nbd & doubtful anyone will notice.

This also updates some data
 * Changes dates to not be so similar
  * It was difficult to tell the order when they're all in the same year
 * Modifies some dummy text for better ledge ability
 * Updates a tag so that the animation will be more fun if you click left to right

Note: Updates .gitignore to remove .vscode dir. I don't think my local vscode
changes are relevant but it is kinda nice to have common format files.

Currently I'm not doing anything weird except using 3 spaces; which I bet
will be a thing at some point, but I'm too used to that to change now and
it's easy enough to auto-fix if the thing becomes a problem.