Support any number of page_index pages
This is a bit of a mess because some DB changes got intermingled
but it shouldn't be too difficult to follow.

Basically, we used to figure out what template to load based on
page_index.slug. We want to add a new page_index.type that will
determine what "type" of template will load.

So we did a few things:
 * Add some general static caching logic to the db libraries
  * I didn't want to copy/paste stuff and it was minor changes
 * Added `page_index.type` to the data source and to each "Page"
 * Created a "PageCollection" to figure out what page should run
   from some given input

NOTE: This doesn't add the db changes b/c I'm not quite sure how
we'll do that.

part of #32