About Me - Abstract out a Widget
We currently only one "Widget" and it's pretty complicated. I'd
like another, more simple demonstration Widget to play with so
this moves the silly "Fun" logic out of the about_me.phtml file
and into it's own BlockOFun Widget.

Then (because Widgets are for now very arbitrary) we can put the
widget back into the about_me page, like we did nothing!

But also:
 * Updated "fun" logic to be more random: random+colors = fun
 * Updated css to be more generic
 * Modified text slightly to (hopefully) explain phtml vs widget

Part of #32