WidgetCollection - Create with WidgetCollectionPacket
Ok, why do we need ANOTHER abstration you ask? Why not use
the existing ArticlePacket?!?

Basically when working with these systems that can pass around and
process huge amounts of data, we want to be able to have easy
POTENTIAL access to data but only access the data we NEED.

This helps us keep track of what systems need what data.

In this instance, the widgets COULD have used ANY data that would
used in the Article but now we know it only uses a Pageid and Type.

We could (in theory) change it to easily not need the ArticlePacket and
we know the only data we'd need to send it was a `pageid` and `type`.

If we had ALL the data, we'd have to verify each Widget didn't use some
random piece of ArticlePacket data. Simple now with the basic Widgets
but difficult if there were 10 of them, each with complicated logic like
the `MiniArticleList` or worse.

So yeah, this creates the WidgetCollectionPacket from an ArticlePacket
and uses that as the source of Widget data during the rendering process.

This also lets us lock down the Widget constructor to protected
 * Nice b/c we have a common static constructor for all the widgets
 * We can call that constructor in any Article template b/c the WCPacket
   can be accessed by the Article object.
  * I.E. we can call any widget in any most template
 * Protected constructors let us do some common logic before/after creating
   the object in the calling static scope

part of #55