ArticlePacket - Rename articleX to pageTypeX
We used to load page templates off of the slug which was attached
to a specific page article so we called the "article templates"

Now those values are based off of a "page type" so calling it
"Article thing" is a bit confusing.

So this updates all the `articleX` references to `pageTypeX` in hope
for better clarity on what's happening here.

Also moves the template rendering logic into the Article which I
think makes it clearer what the scope of the "page type template"
files are.

part of #55