Fix timing on color changes
It's nice to have the colors shift gradually and for them to shift
faster as you click more.

This updates the loop and color change timing so it's a little
nicer to use.

It should be very obvious when a "color boom" happens but color
shift after should be obvious but not immidiate.

As more clicks happen, speed will increase to a max, which isn't
overwhelming but also still very fast.