Add max display width logic
Flexbox is nice because it can grow to any size but our site want
to display max a ~1024 width.

Othewise our content gets so sparse everything looks ugly.

This adds some bounds so that if the window grows beyond 1024px
there's just more right side space.

It's a little odd to be justified left but centering has it's
own weird quirks.