Cleanup footer and Page floating
The Page would grow to a max size, but was justified to the left.

This makes it awkward in large screens because it feels like a lot
of unused pixels they paid for.

So we'll float the page in the center of the background, which gives
a nice illusion of using more of the page than you actually are.

The Footer wasn't aligned correctly because the text was different
sizes and it's difficult to get a correct float with 3 items of
very different sizes.

So we hack in a right/left aligned style & do some fancy media
tags to handle the ugly layout at low widths

Definatly a hack for just this case but also looks good and this
is the demo so yeah, gonna go with it.

Also cleans up some bad spacing.

Also does minor cleanup
 * Fixes some invalid spacing