Break JS into multiple async files
We had one single, large file that held all our JS. This is nice
for deploys because there's only a single file to worry about but
difficult to work with b/c there's a lot of stuff going on in
the file.

Splitting up js into different files and not requiring loading
all the files seperatly is a difficult task and there are many
solutions.

This solution uses an async library to load files "as needed" and
to run some function after they've been loaded.

It's main goal is to break up the files but it also limits the
js files we need to download to ones that are needed when they're
needed which is pretty neat.

Once the basic structure is setup it shouldn't be too difficult
to add more Widgets but if the JS ever gets really complex we're
gonna have a bad time.