Add New App Deploy Config
The old app deployed on an expensive flex machine. It worked but
was way, way more than we needed.

This should update the app to deploy on the free machine. It has
very low resources but we don't care about that.

The best thing is that php8.1 is supported now (in beta) so now
the dev php version can match the production and we're all up
to the latest cool one.