Site: Add a favicon and clean up Head logic
Sites expect a favicon and we didn't have one. Not having one
ends up triggering an (generated by us) error when the website
tries to load a place that doesn't exist.

We could not throw that error, but I want to know anytime someone
goes someplace that doesn't exist. So we'll just generate a simple
favicon and use that (also it's cool to have).

It's really easy to generate favicons now:
https://favicon.io/favicon-generator/

StackOverflow helped with the actual html:
https://stackoverflow.com/a/47303623

This ends up doing a bit of cleanup in places:
 * Moves repeated html out of `head.phtml` into being generated from Head
  * Was going to get messy with all the new favicon stuff
 * Adds a robot.txt to ignore us
  * Got crawled which threw errors
  * We had "no index" but not "no crawl". Now it's also "no crawl".
  * We want to be ignored b/c bots will do all sorts of random things
    to a site.
   * We don't want random things b/c it may costs money if it
      exceeds some limit
 * Cleans up app.yaml
  * Adds favicon to static dirs
  * Removes value needed for MySQL (which we don't use anymore)
  * Links the default favicon locaton to the actual spot
   * This was causing errors to show up whenever a page was loaded
     b/c the browser would request an invalid page and we'd log that
     an invalid page was requested.

closes #57