PageNav: Source data from Firestore
Now that we've got Firestore up and running for page_index we've
got to convert all the other tables too.

This adds the page_nav firestore data to the site by abstracting
the hacky PageIndex logic into a more genereic Firestore object.

It's still pretty hacky but will work to get the data to the site.

It'll have to be updated when we start writing data but that'll
be a bit.

Part of #51