MiniArticle: Pull data from Firestore
This is just a basic dev page hack to see what values we're pulling
in from Firestore.

We can't really do much until we get the tags sorted out but this
does the timestamp converting which is nice to figure out early.

Also adds the "null" value if a requested firestore field is missing.

Missing/NULL values are an expected state for `end_date`

Part of #51