Mini Article Tags: Pull data from firestore
Now that we've added the tag firestore data we want to use it!

Again it's a bit weird how we're building these arrays but we'll
fix that later when we start writing values.

NOTE: Added a hack for newlines. The GUI to edit strings in
firestore is really bad and doesn't seem to support them. Also
it's a input with room for ~20 characters and no option to expand.

Part of #51