Page Nav: Update to new firestore schema
Firestore wants to do references different from how the MySQL db
schema was setup.

Specifically we want to have collections in documents instead of
referencing a document to a collection. It will allow us to do
filtering a searching a lot better and when we get to editing
we won't have any weird hoops to jump through.

This does some extra DB work now that we know:
 a) What db structure we're gonna use (Firestore)
 b) We're gonna have an edit interface soon
  * The Google GUI doesn't handle newlines and we need those.

There's also some abstraction of the Firestore logic now that we
know what we want to do with it and removes the last of the MySQL
logic.

NOTE: This removes the schema file. Sadly there isn't a good firestore
replacement but the static structure should work for people that
care. It's a lot of work to deploy at this point anyways.

Part of #51, #39