MiniArticle: Use Firestore for data
Google Firestore really wants more embedded things so we're gonna
go that route.

This requires some finess as we convert from the old to the new.

In this instance we're still grabbing all the mini articles for
all the pages but the new data source doesn't have a pageid anymore
so we have to stuff that in by iterating over the page_index first.

Eventually we'll actually do some queries and not work with all
the things but for now we'll just hack through this bit to get
reading & the basic code running.

This does a lot of hidden work and took way, way too long for the
amount of code written.

Major Functional Changes:
 - New DB
  - Firestore is awesome in that it's cheap and easy
  - Firestore is terrible b/c there are NO ENFORCED TYPES :(
 - Pages (could) support multiple MiniArticles
  - The way firestore is setup it's easier to support multiple than
    limit to one
 - New Structures
  - Things were setup so that one table referred to another root table
    and we used a "pageid" to keep track of what belongs where
  - Now we embed collections in documents (like firestore wants)
   - This is very, very difficult to manage in the Firestore GUI
   - Had to hack in newlines so we could render parsable text properly
 - New Objects
  - Now there's 3
   - `PageLists`  gathers all the article list data
   - `ArticleCollection` that is a specific "article list"
   - `MiniArticle` that is an "article" in the "article list"

Part of #51, #39