Update PageList and PageNav to use FirestoreConverter
There's a nice object we have for converting values and a trait
that lets us access it so we should use it!

This updates PageList and PageNav to not use the Collections logic
and instead use `DBTraig::getRows` which goes through our silly
converting logic.

This also does some cleanup of names and moves some logic around
to hopefully make it clearer whats happening as we iterate through
the collections.

Part of #51