PageIndex: Add theme support
We had 2 different page colors based on the page type but we'd really
like 4 differnt page colors not tied to any page content.

This adds the "theme" propery to `page_index` and uses that to
choose what stylesheet to display.

It also renames the files to be "$color-theme.less" and adds 2 new
themes: green and grey.

There's some re-arranging of files here too but not much actual
"new" work.

closes #62