PageNav: Update to handle Footer links
Originally the PageNav was designed to support either Article links
or Full URLS.

Now our layout want Article links in the Nav component and full
links in the footer.

This updates the PageNav to support either Footer or Article types.

The Article types are displayed in the Nav component and the footer
will be done in the next commit.

This does some tweaks to the PageNav object to be more Object Oriented
now that there's more filtering and data managment.

It also adds a small css hack to fix the layout shift when you hover
over an article link and it bolds.

Part of #63