Footer: Use PageNav and update style
We added the footer links to the page_nav table but the page still
uses the hardcoded values.

This updates Footer to pull the Footer Type PageNav links and to
display them where the hardcoded vaules where.

This also moves the footer to be contained in the "Section" which
matches how the mockup has it.

Closes #63