Design Alpha: Update HTML structures
The new design has the "main" nav on the right, in the "section"
segment.

Our current layout has the nav on the bottom outside the section
element.

This updates the HTML structure so that we can move the nav from
the right to the bottom of the "Section" element using flexbox
when the page shrinks.

It also put's all the page elements in a container that we can keep
a constant width for all the "page" elements but still have a
background to float on.

The header was given a background color just so we can see what
the boundries are.

Minor name changes happened to be more consistant with each other
and I tried to stop using the general "data" logic b/c it's not
very helpful and is just another layer of obfuscation.

Closes #90