Footer: Fix to bottom on "Phone" sizes
We now want to fix the "footer" to the bottom of the page on phone
sizes but act as a normal footer for the other sizes.

This moves the footer outside the "body-content-container" so it
isn't affected by the page margins (it always fill the page).

It also copies the footer HTML to create a hidden "fixed" div that
we unhide on the min sizes.

NOTE: There's a trick to have 2 footers in the HTML so we get
proper padding at the bottom of the page when the user scrolls
to the bottom. Otherwise the "Fixed" content will overlap with
actual Page content.

Closes #91