PageNav: Change "Article" to "Main"
The final design has 2 "navigation" sections, one that appears
on the right and another that is either a footer or fixed depending
on the page size.

We currently call the section that appears on the right "Article"
but that's an old concept and soon it will have all sorts of links.

This changes the term "Article" to "Main" in the code which should
helpfully clarify things.

Part of #92