PageNav: Support sections in Main
The new design has links grouped under a section title. This adds
a "section" field to the `page_nav` db object and uses that in the
HTML to get the layout as designed.

It's still not fully styled but it at least gets things in the
right spots with the right data.

Part of #92