PageNav: Support images in Main Nav
The new design wants to use some images to link out to Github
and LinkedIn.

The current design doesn't really support images but we can hack
them in pretty easily.

At this point we've given up on any real "general" cases and are
hacking things to get as close to the design as possible.

If we get through all the code changes and still want to do more,
we can worry about solving more "general" cases.

This would be basically just:
 * Defining the PageNavSection
   * Adding a orderby
   * Setting the title
   * Either "text" or "image" type

But we can just hack this in for our single case which keeps us from
having to define another table for the single use case.

Closes #92