SectionFooter: Create html structure and show nav text
We have 2 "footers" a mobile and standard one.

The Mobile footer layout was mostly setup already but the new section
footer needed to be structured.

This adds the footer HTML to the bottom of the Section and adds
some basic CSS so it shows up kinda like how it's mocked up.

Closes #93