PageFooter: Pull image source from DB
We now display an image for some of the footer links but didn't
have a way of actually setting the image anywhere.

Now we pull the image from the db.

This also changes the db name from `srcImage` to `src_img` which
is consistant with the other values.

Part of #93