MobileNav: Update links to be circles
For mobile we're just going to have some colored dots.

This is b/c there's not much to the site and we want people to
click on them to explore.

Most of the pages show the important stuff (it's in the Main Nav)
and the rest of the site just looks pretty.

This updates the links to be circles and changes the background
color to match the "green" theme.

Part of #94