MobileNav: Show page being viewed as a smaller circle
Part of the Navigation effect will be letting the user know where
they are in the "Mobile" steps.

The page they're currently viewing will be a smaller circle than
the other pages.

This updates the mobile footer logic to do that. It was mostly
in place already, just passing the values up to the template and
adding the css.

Part of #94