Design Alpha: Support header images
Each page header was designed as a gif to run some animations but
for now we're gonna stick in some pngs.

This updates the header to support a static image.

There were some hoops to jump through to get all the flex resizing
working and in the end the images are a bit off but close enough.

This is just to get some stuff up in the general right spot, in the
right size and we'll adjust later.

Part of #84