Design Alpha: Modify "Dev" to be "Work"
Now that we actually have a design to impliment and we're onto
style and content we don't need a random "dev" page.

It has all the "Mini Article" content associated with it so we'll
just rename it "Work" and remove all the dev stuff to get a page
close to what was designed.

Part of #85