Design Alpha: Update Mini Article tags
The new design has a | separating tags along with some css changes.

This adds the | between tags and updates the css to be more like
what's defined in the design.

Padding and font sizes are still off b/c we're gonna try to make
those more general after we've done all the basic page changes.

Closes #85