Design Alpha: Update "Robots" Page
The "Robot" page doesn't really have much special going on. It's
mostly just wiki rendered text and theme color changes.

The most work was done changing the "purple" theme to "black" b/c
we'll have a "green" theme but not a "purple" in the end.

The css for the article doesn't quite match the design and it'll
probably end up being a custom page in the end (just for the image
css to look pretty as the pages resize) but we'll wait for content
before worrying about that.

Closes #86