Design Alpha: Update "green" theme colors
We now have a design for the "Life" page which uses the "green"
theme.

Most of the work will be in doing the slideshow but the colors are
now defined so we might as well get that fixed up.

This adds all the "green" theme colors so the page displays closer
to what we've got mocked up.

Part of #87