Slideshow: Generate header html in the Object
We're about to make the Header a lot more complex and mixing PHP
with HTML gets a bit dicey.

We've got a kinda abstraction where we generate HTML in PHP with
the HtmlElement object which is messy but at least gets us in a
consistant language for the complex stuff.

This moves all the existing header phtml logic into the Header
Object which should help the next steps of adding multiple images
easier to handle.

Part of #36