Slideshow: Add support for basic header slideshow
This adds a basic "slideshow" to the top of the Life page.

There's a bunch of fancy JS and CSS magic to handle the transition
movement.

This is a very rough draft but it's functional and looks pretty good.

I ended up getting a lot of direction from:
https://usefulangle.com/post/313/css-flex-order-carousel-infinite

but it was pretty heavily modified to handle mobile and full width
in the header element.

Part of #36