Slideshow: Pull values from the DB
We currently have a slideshow/carousel in the header but the images
are all randomly generated.

We want to pull the image links from the db.

This updates the logic to pull images from a `page_index.page_header_images`
collection instead of just a supporting a single header image.

It's a big weird b/c we have different 2 image sizes to support
for mobile and full width and eventually we'll want a button to
set both to the same image. Also we'll want to support orderby
and maybe alt-text or want to change sizes... anyways we'll want
to support a lot of meta data for each "image" in the slideshow.

This bypasses some of the DBTrait logic b/c we don't really need
to fetch all the page image values.

The fetch logic really should change off of "pageid" but for now
it's how we do stuff & we don't need much more so whatever.

Part of #36