Slideshow: Add basic controller widget
We had a slideshow that was able to go to the "next" slide but
we'd like a "next", "prev" and "set" function.

This requires a lot more fancy JS and CSS work to get looking
pretty.

This is a partial step towards getting things fully working.

The HTML, CSS and JS are setup and there's a DB change that's
needed for ordering but the transitions aren't quite right.

You can go "next", "prev" and set a specific "position" but there
are some bugs and "prev" looks like it's going to "next" when really
we want it to go the other direction.

There's still more JS work to do:
 * abstracting it into it's own file
 * doing some cleanup
 * adding "prev" transition
 * fixing the bugs

but this is mostly functional so we'll commit here and make sure
the existing backing changes are working as intended.

Part of #36