Slideshow: Allow both left and right movement
It's pretty easy to get a slideshow to go right because web layouts
will overflow things to the right and you can use a
`tansform:translateX()` to shift the window and show the things
hidden on the right.

However, if you try to slide left with the same function, you'll see
a lot of empty space b/c the page starts to draw content on the left
side and showing "more left" just shows nothing b/c the page doesn't
know to start drawing before the left side of the page.

The fix is to use a "margin: -100%" on the container which lets the
browser know it should start drawing somthing more on the left side.

Using a bunch of js and css magic we can toggle things to make the
slide work both ways.

Part of #36