Carousel Controller: Add triangle and hover effects
The mockup has triangle images for the left and right carousel
controls, but it's more fun to do SVG polygons b/c then we can
do cool hover effects.

This changes the "Next" and "Prev" actions to have SVG trianges
that turn red on hover.

This also adds a red border active state to the "set" buttons and
when a specific slide is shown, the butten is set as "active"

Part of #36