Life Page: Allow empty text
The layout for the Life Page wanted to have wiki text or else the
width would be too small.

However the mockup just has the Carousel Controller and no text.

This fixes the css so that if there's no text and only the Carousel
Controller, the page will still render full width.

Part of #36