Remove Firestore and convert to static data on Live site
We can't deploy with Firestore anymore because gcloud decided not to
support a needed plugin anymore.

This remove Firestore from composer and the plugin that it needs and
updates the code to pull the static Dev data instead of trying
to use Firestore values.

NOTE: This also removes the Google Ccloud Debugger logic which is
depcrecated.

Closes #100