Design Beta: Animate Welcome Header Circles
Currently all the headers are static and we'd like them to be animated.

The Welcome Header has some circles that we'd like to kinda move around in
the background.

This adds a couple circles and animates them to randomly move to a new X, Y
position over a random time.

It also hacks in some header logic to display the new code.

The JS also got a bit wonky with multiple files requiring the same
external file so the loading order of them got modified a bit.

Part of #108