Design Beta: Animate Welcome Header X Bar
We want the Welcome Header to have cool animations. The Header has
a set of 4 X images in a line that we'd like float kinda slowly in
a random direction before going back to the set 4 position again.

This adds the X png and the html, css and js to do that.

It also alters the Header CSS to be a standard height which I think
is the finalish height we want to use. It also fixes up the slash
image to match the new height.

Part of #108