Design Beta: Add Welcome Header Contact Info
The Welcome Header has my picture and a lot of my contact info.

Now that the background animations are working, we'll add this
static contact data and get it position and looking close to how
we want it in the end. We can fiddle with the final margins and
font sizes when it's all up.

Not a whole lot special here, just adding the HTML & CSS to do the
thing.

Part of #108