Design Beta: Animate Welcome Header Contact Info
The Welcome Header has some contact info that has an animation of
a white rectange moving over the text area and revealing the text
as it moves off of it.

Kinda makes a left to right flash. It's a bit hard to describe but
this does the html, css and js to animate it.

There's nothing to scientific about it, just fiddled with hardcoded
values until I got something that looked like how I want.

Part of #108