Design Beta: Fix Welcome Header Page Sizes
The Welcome Header has some cool animations but they kinda get in the
way of text at certain page sizes.

We have 3 different page sizes & we have more area to cover with the
larger page sizes. This moves the animations into more central
positions and the size of the image depending on the "page size".

It was mostly just fiddling with positions until I got something I
liked.

The random '%' and `vw` are basically so they look cooler when you
resize the window, not much reason other than I liked it.

This also sets the min page size to 375px which is the more common
iPhone mobile size. Could go lower but would have start resizing
text and other page stuff.

Part of #108