Design Beta: Fix Welcome Header Animation in Safari
For some reason Safari didn't like how we set values in js.

This moves all the Welcome Header reveal CSS into the less sheet
and uses classes to manage things instead of just pure js.

It seemed to make Safari happy and is the "proper" way of doing
things.

Part of #108