Design Beta: Animate when user clicks portrait
I thought it would be fun if something happened when the user clicked
on my portrait.

Generally once we animate loading the welcome header text, we don't
reallyhide it again.

However, it's a cool animation to load it all in from the "empty" state
so now when the user clicks on the portrait we hide all the text and
animate like it was a "fresh" page load.

Note: I also noticed the fixed footer z-index was way off so that got fixed.

Part of #108